@@ -72,24 +73,26 @@
     "@types/node": "4.2.20",
     "@types/chai": "^4.0.1",
     "@types/mocha": "^2.2.41",
-    "@types/underscore": "1.8.3",
-    "@types/q": "1.0.5",
-    "@types/invariant": "2.2.29",
-    "@types/marked": "0.3.0",
-    "@types/mkdirp": "0.5.1",
-    "@types/pluralize": "0.0.27",
-    "mocha": "^2.2.1",
-    "chai": "^2.2.0",
+    "@types/underscore": "^1.8.3",
+    "@types/q": "^1.0.5",
+    "@types/invariant": "^2.2.29",
+    "@types/marked": "^0.3.0",
+    "@types/mkdirp": "^0.5.1",
+    "@types/pluralize": "^0.0.27",
+    "@types/fs-extra": "^4.0.2",
+    "mocha": "^4.0.0",
+    "chai": "^4.1.2",
+    "ajv": "^5.2.3",
     "typescript": "2.5.2",
-    "webpack": "1.4.2",
+    "webpack": "^1.15.0",
     "gulp": "^3.9.1",
-    "gulp-mocha": "^2.2.0",
-    "chai-as-promised": "^5.2.0",
-    "json-loader": "^0.5.1",
-    "dev-env-installer": "0.0.5",
+    "gulp-mocha": "^4.3.1",
+    "chai-as-promised": "^7.1.1",
+    "json-loader": "^0.5.7",
+    "dev-env-installer": "^0.0.14",
     "atom-text-typer" : "0.0.1",
     "rimraf": "*",
-    "raml-1-parser-test-utils": "^1.0.7",
+    "raml-1-parser-test-utils": "^1.0.10",
     "gulp-istanbul" : "*"
   },
   "browser": {
